Revert ID column types from `int` to `long` again

+++ src/main/java/org/dependencytrack/model/Tag.java
@@ -53,7 +53,7 @@ public class Tag implements Serializable {
     @PrimaryKey
     @Persistent(valueStrategy = IdGeneratorStrategy.NATIVE)
     @JsonIgnore
-    private int id;
+    private long id;
 
     @Persistent
     @Column(name = "NAME", allowsNull = "false")
@@ -72,7 +72,7 @@ public class Tag implements Serializable {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
